requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
click==8.1.7
colorama==0.4.6
diff-match-patch==20230430
Django==5.0.6
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-import-export==4.0.8
django-redis==5.4.0
django-rest-passwordreset==1.4.1
django-rq==2.10.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.1
pillow==10.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.5
rq==1.16.2
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
